refactor: namespace using fix
