+++ src/Services/Masa.Dcc.Service/Domain/App/Repositories/IBizConfigObjectRepository.cs
@@ -10,6 +10,6 @@ namespace Masa.Dcc.Service.Admin.Domain.App.Repositories
         Task<List<BizConfigObject>> GetListByBizConfigIdAsync(int bizConfigId);
 
         Task<List<(int projectId, ConfigObjectRelease release)>> GetProjectLatestReleaseConfigAsync(
-            List<BuildingBlocks.StackSdks.Pm.Model.ProjectModel> projects, int? envClusterId = null);
+            List<ProjectModel> projects, int? envClusterId = null);
     }
 }
