+++ src/Services/Masa.Dcc.Service/Infrastructure/Repositories/App/BizConfigObjectRepository.cs
@@ -34,7 +34,7 @@ namespace Masa.Dcc.Service.Admin.Infrastructure.Repositories.App
 
         public async Task<List<(int projectId, ConfigObjectRelease release)>>
             GetProjectLatestReleaseConfigAsync(
-                List<BuildingBlocks.StackSdks.Pm.Model.ProjectModel> projects, int? envClusterId = null)
+                List<ProjectModel> projects, int? envClusterId = null)
         {
 
             List<(int projectId, ConfigObjectRelease release)> result = new();
