+++ src/Services/Masa.Dcc.Service/Services/AppService.cs
@@ -1,11 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the Apache License. See LICENSE.txt in the project root for license information.
-using Masa.BuildingBlocks.StackSdks.Auth;
-using Masa.Contrib.StackSdks.Pm;
-using Masa.Dcc.Contracts.Admin.App.Dtos;
-using Masa.Dcc.Service.Admin;
 
-namespace Masa.Dcc.Service.Services;
+namespace Masa.Dcc.Service.Admin.Services;
 
 public class AppService : ServiceBase
 {
