+++ src/Services/Masa.Dcc.Service/_Imports.cs
@@ -27,7 +27,6 @@ global using Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.Queries;
 global using Masa.BuildingBlocks.StackSdks.Auth;
 global using Masa.BuildingBlocks.StackSdks.Auth.Contracts;
 global using Masa.BuildingBlocks.StackSdks.Auth.Contracts.Consts;
-global using Masa.BuildingBlocks.StackSdks.Auth.Contracts.Model;
 global using Masa.BuildingBlocks.StackSdks.Config;
 global using Masa.BuildingBlocks.StackSdks.Dcc.Contracts.Model;
 global using Masa.BuildingBlocks.StackSdks.Middleware;
